Separate getRemoteGitCommit to a separate package
It will be shared between the commit updater and the API server.
