+++ package.json
@@ -18,7 +18,7 @@
     "lint": "eslint *.ts *.mjs && prettier . --check && tsx for_each_package.ts run lint",
     "format": "eslint --fix *.ts *.mjs && prettier . --write && tsx for_each_package.ts run format",
     "compile": "tsc -p tsconfig.json && tsx for_each_package.ts run compile",
-    "dev": "npm run build && npm run start_test_containers && ./scripts/init_test_db.sh && node dist/index.js",
+    "dev": "npm run build && npm run start_test_containers && ./scripts/init_test_db.sh && node dist/api-server/index.js",
     "postinstall": "tsx for_each_package.ts install"
   },
   "devDependencies": {
