+++ packages/api-server/Dockerfile
@@ -19,10 +19,11 @@
 # Build the app -------------
 FROM docker.io/node:22.13.1-alpine3.21@sha256:e2b39f7b64281324929257d0f8004fb6cb4bf0fdfb9aa8cedb235a766aec31da as builder
 
-WORKDIR /app
+WORKDIR /api-server
 
 COPY . ./
 COPY --from=root prisma/ ./prisma/
+COPY --from=root packages/common/ /common
 RUN npm install -g npm && npm ci && npm run build && rm dist/*.tsbuildinfo
 
 # Production image --------------
@@ -32,7 +33,7 @@ WORKDIR /app
 LABEL org.opencontainers.image.authors="8 Hobbies, LLC"
 LABEL org.opencontainers.image.licenses=AGPL-3.0-or-later
 
-RUN apt-get update && apt-get install --no-install-recommends -y git-core && rm -rf /var/lib/apt/lists/*
+RUN apt-get update && apt-get install --no-install-recommends -y git-core
 
 ENV NODE_ENV=production
 
@@ -40,7 +41,7 @@ RUN addgroup --system --gid 1001 nodejs
 RUN adduser --system --uid 1001 nodejs
 
 COPY package.json package-lock.json ./
-COPY --from=builder /app/dist/ ./
+COPY --from=builder /api-server/dist/ ./
 RUN npm install -g npm && npm ci
 
 USER nodejs
@@ -50,4 +51,4 @@ EXPOSE 3000
 ENV PORT=3000
 ENV HOSTNAME="::"
 
-ENTRYPOINT ["node", "index.js"]
+ENTRYPOINT ["node", "api-server/src/index.js"]
