+++ packages/common/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "commitrack Common",
+  "version": "1.0.0",
+  "private": true,
+  "description": "Commit Tracker",
+  "license": "AGPL-3.0-or-later",
+  "author": "8 Hobbies, LLC",
+  "type": "module",
+  "scripts": {
+    "lint": "eslint src/*.ts *.mjs && prettier . --check",
+    "build": "npm run compile",
+    "build_container_img": "echo 'no container image for common'",
+    "format": "eslint --fix src/*.ts *.mjs && prettier . --write",
+    "compile": "tsc -p tsconfig.prod.json"
+  },
+  "devDependencies": {
+    "@8hobbies/eslint-conf-baseline": "^6.0.0",
+    "@types/eslint": "^9.6.1",
+    "@types/eslint__js": "^8.42.3",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.10.7",
+    "@vitest/coverage-v8": "^3.0.3",
+    "eslint": "^9.18.0",
+    "globals": "^15.14.0",
+    "prettier": "^3.4.2",
+    "prisma": "^6.2.1",
+    "typescript": "~5.7.3",
+    "typescript-eslint": "^8.21.0"
+  }
+}
