+++ packages/common/tsconfig.base.json
@@ -1,0 +1,21 @@
+{
+  "compilerOptions": {
+    "target": "ESNext",
+    "lib": ["esnext"],
+    "allowJs": true,
+    "outDir": "./dist",
+    "skipLibCheck": true,
+    "strict": true,
+    "esModuleInterop": true,
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "jsx": "preserve",
+    "incremental": true,
+    // Required by @fastify/type-provider-json-schema-to-ts
+    "noStrictGenericChecks": false
+  },
+  "include": ["**/*.ts"],
+  "exclude": ["node_modules", "dist/", "docs/"]
+}
