+++ packages/common/tsconfig.json
@@ -1,0 +1,11 @@
+// Only for better editor/IDE experience
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "noEmit": true,
+    "allowImportingTsExtensions": true,
+    "types": ["vitest/globals"]
+  },
+  "include": ["**/*.ts"],
+  "exclude": ["node_modules", "dist/", "docs/"]
+}
