+++ packages/common/tsconfig.prod.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.base.json",
+  "include": ["src/*.ts"],
+  "exclude": ["node_modules", "dist/", "docs/", "**/*.test.ts"]
+}
